requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
starlette==0.27.0

# Image processing and computer vision
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.26.2
scikit-image==0.25.2
tifffile==2025.5.10

# Machine Learning - PyTorch ecosystem
torch==2.7.1
torchvision==0.22.1
efficientnet-pytorch==0.7.1
pytorch-metric-learning==2.8.1

# Machine Learning - TensorFlow (for model training)
tensorflow>=2.0.0

# Machine Learning - Additional ML libraries
scikit-learn==1.7.0
scipy==1.15.3

# Geometry processing
Shapely==2.0.2
pyclipper==1.3.0.post6

# Cloud storage and file handling
cloudinary==1.44.1
boto3==1.38.44
botocore==1.38.44
fsspec==2025.5.1

# HTTP and networking
requests==2.32.4
httpx==0.28.1
httpcore==1.0.9
httptools==0.6.4

# Data processing and utilities
pandas>=1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1

# Security and authentication
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# File and data handling
aiofiles==24.1.0
pathlib2>=2.0.0
py7zr==1.0.0
rarfile==4.2
zipfile36>=0.1.0

# JSON and data serialization
simplejson==3.20.1
PyYAML==5.1

# Logging and monitoring
python-json-logger==2.0.7
logging>=0.4.9.6

# Testing and development
pytest==7.4.3
httpx==0.25.2

# Additional utilities
tqdm==4.67.1
click==8.2.1
colorama==0.4.6
retrying==1.4.0
tabulate==0.9.0

# Optional: Visualization (if needed)
matplotlib>=3.0.0
seaborn==0.13.2

# Optional: Database (if needed)
pymongo==4.9.2

# Optional: Additional ML tools
fiftyone==1.6.0
fiftyone-brain==0.21.2
fiftyone_db==1.1.7
pycocotools==2.0.10 